chore: fix clippy warnings, add Cargo.lock, update test counts
- NFS: remove unnecessary u64 casts, allow dead_code on stored fields
- FUSE: remove unnecessary u32 casts in statfs
- Add src/nfs/Cargo.lock (required for binary crates)
- HANDOVER.md: 52 Rust + 28 Python = 80 tests total

+++ src/fuse/fuse-driver/src/passthrough.rs
@@ -518,9 +518,9 @@ impl Filesystem for IcloudFs {
                     stat.f_bavail,
                     stat.f_files,
                     stat.f_ffree,
-                    stat.f_bsize as u32,
+                    stat.f_bsize,
                     255, // namelen
-                    stat.f_bsize as u32, // frsize
+                    stat.f_bsize, // frsize
                 );
             } else {
                 reply.error(Errno::EIO);
